+++ src/components/car/CarTable.vue
@@ -1,12 +1,13 @@
 <template>
   <div>
-    <el-table :data="cars" height="400" border style="width: 100%" ref="user-table">
-      <el-table-column label="车辆识别码" prop="vin" sortable width="170"></el-table-column>
+    <el-table :data="cars" border style="width: 100%;height:100%" ref="user-table">
+      <el-table-column label="ID" prop="id" sortable width="170"></el-table-column>
+      <el-table-column label="VIN码" prop="vin" sortable width="170"></el-table-column>
       <el-table-column label="车牌号" prop="plate" sortable width="170"></el-table-column>
       <el-table-column label="车辆类型" prop="vehicleType" sortable width="170"></el-table-column>
       <el-table-column label="所有人" prop="owner" sortable width="170"></el-table-column>
       <el-table-column label="发动机号" prop="engine" sortable width="170"></el-table-column>
-      <el-table-column label="车型" prop="model" sortable width="170"></el-table-column>
+      <el-table-column label="车型" prop="model" sortable></el-table-column>
       <el-table-column label="操作" width="160">
         <template slot-scope="scope">
           <el-button size="mini" @click="edit(scope.$index, scope.row)">编辑</el-button>
@@ -38,6 +39,7 @@
       }
     },
     mounted() {
+      this.params.owner = this.$route.query.userId;
       this.get();
       this.$Bus.$on('carQueryEvent', (queryParams) => {
         this.params = queryParams;
@@ -64,9 +66,26 @@
         this.$Bus.$emit('carEditEvent', item);
       },
       deleteItem(index, item) {
-        this.$http.delete('/api/car/', item.id, () => {
-          this.get();
+        this.$confirm('此操作将永久删除该记录, 是否继续?', '提示', {
+          confirmButtonText: '确定',
+          cancelButtonText: '取消',
+          type: 'warning'
         })
+          .then(() => {
+            this.$http.delete('/api/car/', item.id, () => {
+              this.$message({
+                type: 'success',
+                message: '删除成功!'
+              });
+              this.get();
+            });
+          })
+          .catch(() => {
+            this.$message({
+              type: 'info',
+              message: '已取消删除'
+            });
+          })
       }
     }
   }
